requirements.txt
fastapi
uvicorn
requests
python-dotenv
redis
langchain
langchain-groq
langchain-huggingface
sentence-transformers
faiss-cpu
